Move EndpointPlugin class definition to plugins
This aligns with the patterns of other plugins like IOProcessor.
vllm/entrypoints/serve/endpoint_plugin.py -> vllm/plugins/endpoint_plugins/interface.py

Review comment:
- https://github.com/vllm-project/vllm/pull/47454#discussion_r3517322953
